feat(v3.6.0): Phase 5 - Congés Décimal + Réorganisation Paramètres UI
- Backend: Retour API congés en float (jours_conges_acquis/pris/restants)
- Frontend: Affichage décimal .toFixed(2) dans CongesList
- Frontend: Réorganisation ParametresSalaires en sections visuelles:
   INDEMNITÉS (IN, IFSP, IEP)
   PRIMES (Encouragement, Chauffeur, Nuit, Femme Foyer, Panier, Transport)
   RETENUES (Sécu Sociale)
   PARAMÈTRES CALCUL (Congés, Options avancées)
- UI: Labels en gras, bordures colorées par section, dividers

+++ backend/routers/conges.py
@@ -75,9 +75,9 @@ def list_conges(
             annee=c.annee,
             mois=c.mois,
             jours_travailles=c.jours_travailles,
-            jours_conges_acquis=int(c.jours_conges_acquis or 0),
-            jours_conges_pris=int(c.jours_conges_pris or 0),
-            jours_conges_restants=int(c.jours_conges_restants or 0)
+            jours_conges_acquis=float(c.jours_conges_acquis or 0),  # ⭐ v3.6.0 Phase 5: Décimal
+            jours_conges_pris=float(c.jours_conges_pris or 0),  # ⭐ v3.6.0
+            jours_conges_restants=float(c.jours_conges_restants or 0)  # ⭐ v3.6.0
         ))
         
     return results
